[MRG] Automatic merge, trunk 24715 to 24733
git-svn-id: https://svn.code.sf.net/p/tikiwiki/code/branches/experimental@24737 b456876b-0849-0410-b77d-98878d47e9d5

+++ tiki-log-merge/lib/test/auth/AuthTokensTest.php
@@ -1,19 +1,26 @@
 <?php
 
+/** 
+ * @group integration
+ */
+
 require_once 'lib/auth/tokens.php';
 
 class AuthTokensTest extends PHPUnit_Framework_TestCase
 {
 	private $db;
+	
 	function setUp() {
 		$this->db = TikiDb::get();
 		$this->db->query( 'TRUNCATE tiki_auth_tokens' );
 	}
 
 	function tearDown() {
-		$this->db->query( 'TRUNCATE tiki_auth_tokens' );
+		if ($this->db) {
+			$this->db->query( 'TRUNCATE tiki_auth_tokens' );
+		}
 	}
-	
+		
 	function testNoTokensIsDenied() {
 		$lib = new AuthTokens( $this->db );
